fastmcp-solution/requirements.txt
# FastMCP 解决方案 - 使用兼容的版本
fastmcp>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0